requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
diff-match-patch==20241021
Django==4.2.16
django-admin-interface==0.29.1
django-admin-numeric-filter==0.1.9
django-admin-sortable2==2.2.3
django-appconf==1.0.6
django-auditlog==3.0.0
django-ckeditor==6.4.2
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-dynamic-raw-id==4.3
django-filter==24.3
django-imagekit==5.0.0
django-import-export==4.1.1
django-jalali-date==1.1.3
django-js-asset==2.2.0
django-json-widget==2.0.1
django-modeltranslation==0.18.11
django-mptt==0.14.0
django-mptt-urls==2.0.4
django-phonenumber-field==8.0.0
django-ranged-response==0.2.0
django-ratelimit==3.0.1
django-simple-captcha==0.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.10
importlib-resources==6.4.5
inflection==0.5.1
isodate==0.7.2
jalali-core==1.0.0
jdatetime==5.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.3.0
phonenumbers==8.13.49
pilkit==3.0
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
psycopg==3.2.3
psycopg2-binary==2.9.10
pycryptodome==3.21.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rpds-py==0.20.1
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tablib==3.5.0
text-unidecode==1.3
typing-extensions==4.12.2
tzdata==2024.2
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
user-agents==2.2.0
xmltodict==0.14.2
zeep==4.3.1
